requirements.txt
https://pypi.org/simple
jupyter==*
torch==*
torchvision==*
matplotlib==*
scikit-learn==*
tqdm==*
ray[tune]==*
pytorch-lightning==*
gym==*
gymnasium==*
tree==*
dm-tree==*
scikit-optimize==*
bayesian-optimization==*
botorch==*
seaborn==*
optuna==*
shap==*
flask==*
pillow==*
tkinterdnd2==*
customtkinter==*
